+++ src/_utils/dom/index.ts
@@ -1,1 +1,2 @@
 export { isDocument } from './is-document'
+export { download } from './download'
