+++ src/data-table/src/interface.ts
@@ -515,7 +515,7 @@ export interface DataTableInst {
   page: (page: number) => void
   sort: (columnKey: ColumnKey, order: SortOrder) => void
   scrollTo: ScrollTo
-  exportCsv: (options?: CsvOptionsType) => void
+  downloadCsv: (options?: CsvOptionsType) => void
   /** @deprecated it but just leave it here, it does no harm */
   clearFilter: () => void
 }
